Better name refactoring
